blocking tangram every time
